+++ devicemanagement/conanfile.py
@@ -36,7 +36,7 @@ class DeviceManagementConan(ConanFile):
         self.version = get_version(self, self.name)
 
     def requirements(self):
-        self.requires("deviceApi/[>=1.4.0 <2.0.0]")
+        self.requires("deviceApi/2.0.0-alpha")
         self.requires("deviceLayer/[>=2.1.0 <3.0.0]")
         self.requires("hostUtils/[>=0.2.0 <1.0.0]")